+++ source/Nuke.Common.Tests/ArgumentsTest.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
 using Nuke.Common.Tooling;
+using Nuke.Common.Utilities.Collections;
 using Xunit;
 
 namespace Nuke.Common.Tests
