+++ source/Nuke.Common/Utilities/Assembly.Versioning.cs
