+++ source/Nuke.Common/Utilities/Disposable.CombineWith.cs
